backend/requirements.txt
# Python dependencies
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.11
python-dotenv==1.0.0
pandas==2.0.3
matplotlib==3.7.1
statsmodels==0.14.0
jinja2==3.1.2
xhtml2pdf==0.2.7
plotly==5.15.0
google-generativeai==0.1.0  # check for latest version
